+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # 1.0.1
 
-- remove scaffold folder when removing config from gt
+- added `gt config list`
+- removed scaffold folder when removing config from gt
+- silent output when getting git repo url
 
 # 1.0.0
 
-add config command
+added config command
 
 # 0.0.0